requirements.txt
google-generativeai == 0.3.*
Pillow == 10.*

